+++ app/src/main/java/Entities/ItemDisplays/RecipeItemDisplay.java
@@ -5,11 +5,11 @@ import Entities.Ingredient;
 public interface RecipeItemDisplay {
     /**
      * Returns display information for each respective recipe item type and quantity
-     * @param quantity: float representing quantity
+     * @param quantity: Double representing quantity
      * @param ingredient: ingredient that the recipeItem is
      * @return String representing display
      */
-    String display(float quantity, Ingredient ingredient);
+    String display(double quantity, Ingredient ingredient);
 
     /**
      * Return serialized type for RecipeItem
